+++ engine/sqlmanager/src/index.js
@@ -18,10 +18,36 @@
 module.exports = async (...args) => {
   return new Promise(async (resolve, reject) => {
     const [params, obj] = args;
-
+    const [pathname, curdir] = params;
+    const [library, sys, cosetting] = obj;
+    const bcrypt = require("bcrypt");
+    const csv = require("csv-parser");
+    const {
+      utils: { handler },
+    } = library;
     try {
       let lib = {
         sqlite3: await require("./sqlite3")(params, obj),
+      };
+
+      lib["password"] = (...args) => {
+        let [password, hashpassword] = args;
+        let output = handler.dataformat;
+        try {
+          let saltrounds = 10;
+          if (password && hashpassword) {
+            let rtn = bcrypt.compareSync(password, hashpassword);
+            output.data = { result: rtn, status: "compare" };
+          } else {
+            let rtn = bcrypt.hashSync(password, saltrounds);
+            output.data = { result: rtn, status: "generate" };
+          }
+        } catch (error) {
+          output = errhandler(error);
+          errlog(error);
+        } finally {
+          return output;
+        }
       };
 
       resolve(lib);